pyproject.toml
[project]
name = "multi-alignement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "biopython>=1.85",
    "counter>=1.0.0",
    "ete3>=3.1.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "scipy>=1.16.1",
]
